# backend/cmd/api/docs/swagger.yaml
basePath: /api
definitions:
  github_com_NeroQue_course-management-backend_internal_models.CreateCourseInput:
    properties:
      base_path:
        type: string
      creator:
        type: string
      creator_id:
        type: string
      description:
        type: string
      relative_path:
        type: string
      title:
        type: string
    type: object
...
  internal_api_handlers.BatchImportRequest:
    properties:
      courses:
        items:
          $ref: '#/definitions/github_com_NeroQue_course-management-backend_internal_models.CreateCourseInput'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing courses, modules, profiles, and user progress in a
    learning management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Course Management System API
  version: "1.0"
paths:
...
  /content/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a specific content item as completed for a user
      parameters:
      - description: Content Item ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          properties:
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Content marked as completed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Mark content as completed
      tags:
      - progress
  /content/{id}/progress:
    post:
      consumes:
      - application/json
      description: Update progress information for a specific content item
      parameters:
      - description: Content Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Progress update data
        in: body
        name: progress
        required: true
        schema:
          properties:
            completed:
              type: boolean
            last_position:
              type: number
            progress_pct:
              type: number
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Progress updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing fields
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update content progress
      tags:
      - progress
  /courses:
    get:
      consumes:
      - application/json
      description: Get a list of all available courses
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved courses
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Import a new course from a directory path
      parameters:
      - description: Course creation input with title and relative_path
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/github_com_NeroQue_course-management-backend_internal_models.CreateCourseInput'
      produces:
      - application/json
      responses:
        "201":
          description: Course created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing required fields
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User must be logged in
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionAuth: []
      summary: Create a new course
      tags:
      - courses
  /courses/{id}/progress:
    get:
      consumes:
      - application/json
      description: Get progress information for a specific course and user
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course progress information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid course or user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get course progress
      tags:
      - progress
  /courses/batch:
    post:
      consumes:
      - application/json
      description: Import multiple courses at once from provided course inputs
      parameters:
      - description: Batch import request with array of courses
        in: body
        name: batchRequest
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.BatchImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns task ID for tracking import progress
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or empty course list
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User must be logged in
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionAuth: []
      summary: Batch import courses
      tags:
      - courses
  /courses/directories:
    get:
      consumes:
      - application/json
      description: Get a list of all available course directories in the filesystem
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved directories
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List course directories
      tags:
      - courses
  /courses/scan:
    get:
      consumes:
      - application/json
      description: Find course directories that haven't been imported yet
      produces:
      - application/json
      responses:
        "200":
          description: Successfully scanned for new courses
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Scan for new courses
      tags:
      - courses
  /modules/{id}/progress:
    get:
      consumes:
      - application/json
      description: Get progress information for a specific module and user
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Module progress information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid module or user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get module progress
      tags:
      - progress
...
  /users/{id}/progress:
    get:
      consumes:
      - application/json
      description: Get overall progress summary for a user across all courses
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User progress summary
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user progress summary
      tags:
      - progress
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"